feat: local-volatility MCL diffusion (SABR -> Dupire) for single-asset
Wires the previously-dead LocalVolatilityNode into the MCL spot diffusion so a
SABR (local) surface is Monte-Carlo diffused as a full Dupire local-vol surface,
instead of throwing "use the PDE pricer". Single-asset (mono) only for now.

- Single::GetNode branches on _volatility->_is_local: a local surface builds a
  LocalVolatilityNode (via new BuildLocalVolNode), a flat surface keeps the
  constant-vol node. BuildLocalVolNode samples Equity::GetLocalVolatility (Dupire)
  onto a per-diffusion-date log-spot grid sized like the PDE (+/- 8*ATM-vol*sqrt(T),
  201 points), built through the scenario-suffixed GetOrCreate so vega/rho/spot
  bumps re-sample it.
- LocalVolatilityNode: ComputeValue now does clamped linear interpolation on the
  grid (was an unclamped nearest-neighbour read -> out-of-range UB on tail paths)
  and guards DateIndex 0; GetDateDependencies (was empty) now declares the spot
  dependency at DateIndex-1 so the topo-sort orders the path before the lookup;
  destructor frees the owned vol vectors; offset is signed (spot < 1 / FX).
- NodeCollector::GetDateList exposes the diffusion schedule to node builders.

Test (Dupire repricing property): a European call under local-vol MCL matches the
ANA implied-vol price within MC error, both for a flat SABR surface (constant local
vol) and a genuine smile (rho=-0.3, nu=0.4) — agreement ~0.05 on a ~14 premium.
README limitation updated. Milstein step follows next.

+++ src/nodes/local_volatility_node.hpp
@@ -5,15 +5,14 @@ class LocalVolatilityNode : public MonteCarloNode
 {
 
   private:
-    //! gsl interpolation stuff
-    // gsl_interp_accel * _g_acc;
-    // gsl_spline * _g_spline;
-
+    //! per diffusion date: the local vol sampled on a log-spot grid, the grid
+    //! spacing in log-spot, and the (signed) log-spot index of the grid's first
+    //! point (so grid point i sits at log-spot = (offset + i) * ln_step).
     vector<la_vector*> _vol_vector_list;
     vector<double> _ln_step_list;
-    vector<size_t> _offset_list;
+    vector<long> _offset_list;
 
-    //! for fast interpolation
+    //! the spot path this surface is sampled along (reads the previous step)
     MonteCarloNode* _spot_node;
 
   public:
@@ -24,7 +23,7 @@ class LocalVolatilityNode : public MonteCarloNode
 
     void SetSpotNode( MonteCarloNode* SpotNode );
     void PushLnStep( double LnStep );
-    void PushOffset( size_t Offset );
+    void PushOffset( long Offset );
     void PushVolVector( la_vector* VolVector );
 
     LocalVolatilityNode( const string& Name );
